mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "tominmin", email = "fate59133@gmail.com" }
]
dependencies = [
    "mcp>=1.7.1",
    "anthropic>=0.50.0",
    "python-dotenv>=1.1.0",
    "openai>=1.77.0",
]
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
